pyproject.toml
[project]
name = "oss-analyzer"
version = "0.1.0"
description = " ff"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "sh>=2.2.2",
]

[dependency-groups]
dev = [
    "marimo>=0.14.13",
    "pandas>=2.3.1",
]
